Relocate cue registration function to Application
