+++ lisp/plugins/misc_cues/__init__.py
@@ -21,8 +21,6 @@ from os import path
 
 from lisp.core.loading import load_classes
 from lisp.core.plugin import Plugin
-from lisp.cues.cue_factory import CueFactory
-from lisp.ui.ui_utils import translate
 
 
 class MiscCues(Plugin):
@@ -35,4 +33,4 @@ class MiscCues(Plugin):
 
         # Register the action-cue in the cue-factory
         for _, cue in load_classes(__package__, path.dirname(__file__)):
-            CueFactory.register_cue_type(self.app, cue)
+            app.register_cue_type(cue)
