+++ lisp/plugins/osc/osc.py
@@ -18,7 +18,6 @@
 
 
 from lisp.core.plugin import Plugin
-from lisp.cues.cue_factory import CueFactory
 from lisp.plugins.osc.osc_cue import OscCue
 from lisp.plugins.osc.osc_server import OscServer
 from lisp.plugins.osc.osc_settings import OscSettings
@@ -43,7 +42,7 @@ class Osc(Plugin):
         )
 
         # Register the OSC cue type
-        CueFactory.register_cue_type(self.app, OscCue, translate("CueCategory", "Protocol cues"))
+        app.register_cue_type(OscCue, translate("CueCategory", "Protocol cues"))
 
         # Create a server instance
         self.__server = OscServer(
